feat(api): Rename `errors` to `exceptions`
The `errors` module got renamed to `exceptions`, since the latter is
more appropriate.
