--- docs/api/errors.rst
@@ -1,5 +1,0 @@
-Errors
-------
-
-.. automodule:: piquasso.api.errors
-   :members:
