+++ docs/api/exceptions.rst
@@ -1,0 +1,5 @@
+Exceptions
+----------
+
+.. automodule:: piquasso.api.exceptions
+   :members:
