+++ piquasso/api/exceptions.py
