+++ tests/instructions/test_preparations.py
@@ -23,5 +23,5 @@ def test_StateVector_is_valid_specifying_floats_close_to_integers():
 
 
 def test_StateVector_raises_InvalidState_when_nonintegers_specified():
-    with pytest.raises(pq.api.errors.InvalidState):
+    with pytest.raises(pq.api.exceptions.InvalidState):
         pq.StateVector([1, 1.3, 2.4])
